a function that reads a text file and prints it to the POSIX standard output

+++ 0x15-file_io/0-read_textfile.c
@@ -8,7 +8,7 @@
  */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	ssize_t op, re, wr;
+	int op, re, wr;
 	char *buffer;
 
 	if (filename == NULL)
